requirements.txt
# A custom version of transformers for the project
-e git+https://github.com/dakinggg/transformers@defbd001072dffbedf03c7f9eb880f2836c640e3#egg=transformers


# Other dependencies
fuzzysearch==0.7.1
textacy==0.10.0
datasets==1.3.0
scispacy
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_sm-0.5.1.tar.gz

# Misc 
fire 